requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
schedule>=1.1.0
requests>=2.26.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # Para PostgreSQL
python-dotenv>=0.19.0

# Performance monitoring
psutil>=5.8.0
plotly>=5.3.0
dash>=2.0.0

# Data analysis and visualization
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0

# API integrations
python-binance>=1.0.15
ccxt>=1.60.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Utilities
tqdm>=4.62.0
loguru>=0.5.0
python-telegram-bot>=13.7  # Para notificaciones 